+++ src/common/appearance/index.js
@@ -1,2 +1,3 @@
 export * from "./colors.js";
 export * from "./fonts.js";
+export * from "./weights.js";