+++ src/common/appearance/weights.js
@@ -1,0 +1,6 @@
+export const weights = [
+    "normal",
+    "bold"
+];
+
+export const defaultWeight = weights[1];