+++ src/common/constants/options.js
@@ -8,4 +8,6 @@ export const RESULTS_PER_PAGE = "results-per-page";
 export const RESULTS_SORTING = "results-sorting";
 
 export const UI_SELECTED_ITEM_COLOR = "ui-selected-item-color";
+export const UI_SELECTED_ITEM_ARROW = "ui-selected-item-arrow";
+export const UI_SELECTED_ITEM_FONT_WEIGHT = "ui-selected-item-font-weight";
 export const UI_FONT_SIZE = "ui-font-size";